requirements.txt
alembic==1.8.1
APScheduler==3.9.1.post1
beautifulsoup4==4.11.1
django-environ==0.9.0
fastapi==0.86.0
latest-user-agents==0.0.3
loguru==0.6.0
pandas==1.5.1
passlib==1.7.4
pytest==7.2.0
python-multipart==0.0.5
requests==2.28.1
sentry-sdk==1.10.1
SQLAlchemy==1.4.41
sqlmodel==0.0.8
tqdm==4.64.1
uvicorn==0.19.0
syspath==3.0.1